+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.66 (not yet released)
 ------------
 
+* Fixed issue with get participant actions.
+
 Version 1.65
 ------------